backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asgiref==3.11.1",
    "django==6.0.3",
    "django-cors-headers==4.9.0",
    "django-filter==25.2",
    "djangorestframework==3.17.1",
    "djangorestframework-simplejwt==5.5.1",
    "drf-yasg==1.21.15",
    "inflection==0.5.1",
    "packaging==26.0",
    "pyjwt==2.12.1",
    "pytz==2026.1.post1",
    "pyyaml==6.0.3",
    "sqlparse==0.5.5",
    "tzdata==2025.3",
    "uritemplate==4.2.0",
]
